Fix the confusion: serious brainfart there.
I was saving my changes in the wrong file.
Another living proof that GHC is almost never wrong; it is often the human.
